+++ src/learning.rs
@@ -118,7 +118,7 @@ mod tests {
 
     fn get_weigths(v: f64) -> NetworkWeights<f64> {
         let mut rn = RoadNetworkWeights::with_capacity(1, 1);
-        rn[0].push(TTF::Constant(Time(v)));
+        rn[0].insert(0, TTF::Constant(Time(v)));
         NetworkWeights::new(Some(rn))
     }
 
@@ -177,14 +177,14 @@ mod tests {
         let model = LearningModel::Exponential(exp_model);
         assert_eq!(model.learn(&w1, &w2, 0), get_weigths(20.));
         let x2 = model.learn(&w1, &w2, 1);
-        if let TTF::Constant(v) = x2.get_road_network().unwrap()[0][0] {
+        if let TTF::Constant(v) = x2.get_road_network().unwrap()[(0, 0)] {
             let expected = (20. + 0.8 * 10.) * 0.2 / (1. - 0.8f64.powi(2));
             assert!((v.0 - expected).abs() < 1e-4, "{:?} != {:?}", v, expected);
         } else {
             panic!("Invalid road network weight: {:?}", x2.get_road_network());
         }
         let x3 = model.learn(&x2, &w3, 2);
-        if let TTF::Constant(v) = x3.get_road_network().unwrap()[0][0] {
+        if let TTF::Constant(v) = x3.get_road_network().unwrap()[(0, 0)] {
             let expected = (30. + 0.8 * 20. + 0.8f64.powi(2) * 10.) * 0.2 / (1. - 0.8f64.powi(3));
             assert!((v.0 - expected).abs() < 1e-4, "{:?} != {:?}", v, expected);
         } else {