+++ src/network/road_network/weights.rs
@@ -6,11 +6,12 @@
 //! Description of the [RoadNetworkWeights].
 use std::ops::{Index, IndexMut};
 
-use petgraph::graph::EdgeIndex;
+use hashbrown::HashMap;
 use schemars::JsonSchema;
 use serde_derive::{Deserialize, Serialize};
 use ttf::{TTFNum, TTF};
 
+use super::OriginalEdgeIndex;
 use crate::units::Time;
 
 /// Structure to store the travel-time functions of each edge of a
@@ -18,23 +19,24 @@ use crate::units::Time;
 ///
 /// The outer vector has the same length as the number of unique vehicles of the associated
 /// [RoadNetwork](super::RoadNetwork).
-/// The inner vectors have all the same length (i.e., the RoadNetworkWeights represent a matrix)
-/// which is equal to the number of edges of the associated [RoadNetwork](super::RoadNetwork).
 #[derive(Clone, Debug, PartialEq, Deserialize, Serialize, JsonSchema)]
 #[serde(bound = "T: TTFNum")]
 #[schemars(title = "Road Network Weights")]
 #[schemars(
-    description = "Travel-time functions of each edge of road network (inner array), for each \
-    unique vehicle (outer array)."
+    description = "Travel-time functions of each edge of road network, for each unique vehicle \
+    (outer array)."
 )]
-pub struct RoadNetworkWeights<T>(Vec<Vec<TTF<Time<T>>>>);
+pub struct RoadNetworkWeights<T>(
+    #[schemars(with = "Vec<std::collections::HashMap<OriginalEdgeIndex, TTF<Time<T>>>>")]
+    Vec<HashMap<OriginalEdgeIndex, TTF<Time<T>>>>,
+);
 
 impl<T> RoadNetworkWeights<T> {
     /// Initializes a new RoadNetworkWeights instance with enough capacity for the given number of
     /// unique vehicles and edges.
     pub fn with_capacity(nb_unique_vehicles: usize, nb_edges: usize) -> Self {
         let mut vehicle_weights = Vec::new();
-        vehicle_weights.resize_with(nb_unique_vehicles, || Vec::with_capacity(nb_edges));
+        vehicle_weights.resize_with(nb_unique_vehicles, || HashMap::with_capacity(nb_edges));
         RoadNetworkWeights(vehicle_weights)
     }
 
@@ -76,10 +78,19 @@ impl<T: TTFNum> RoadNetworkWeights<T> {
         let (nb_vehicles, nb_edges) = self.shape();
         let mut new_weights = RoadNetworkWeights::with_capacity(nb_vehicles, nb_edges);
         for (i, (self_weights, other_weights)) in self.0.iter().zip(other.0.iter()).enumerate() {
-            for (self_ttf, other_ttf) in self_weights.iter().zip(other_weights.iter()) {
-                new_weights.0[i].push(self_ttf.apply(other_ttf, |w1, w2| {
-                    Time(coefficient * w1.0 + (T::one() - coefficient) * w2.0)
-                }));
+            for ((self_id, self_ttf), (other_id, other_ttf)) in
+                self_weights.iter().zip(other_weights.iter())
+            {
+                assert_eq!(
+                    self_id, other_id,
+                    "The weights do not have the same edge ids"
+                );
+                new_weights.0[i].insert(
+                    *self_id,
+                    self_ttf.apply(other_ttf, |w1, w2| {
+                        Time(coefficient * w1.0 + (T::one() - coefficient) * w2.0)
+                    }),
+                );
             }
         }
         new_weights
@@ -102,10 +113,19 @@ impl<T: TTFNum> RoadNetworkWeights<T> {
         let (nb_vehicles, nb_edges) = self.shape();
         let mut new_weights = RoadNetworkWeights::with_capacity(nb_vehicles, nb_edges);
         for (i, (self_weights, other_weights)) in self.0.iter().zip(other.0.iter()).enumerate() {
-            for (self_ttf, other_ttf) in self_weights.iter().zip(other_weights.iter()) {
-                new_weights.0[i].push(self_ttf.apply(other_ttf, |w1, w2| {
-                    Time(w1.0.powf(a / (a + b)) * w2.0.powf(b / (a + b)))
-                }));
+            for ((self_id, self_ttf), (other_id, other_ttf)) in
+                self_weights.iter().zip(other_weights.iter())
+            {
+                assert_eq!(
+                    self_id, other_id,
+                    "The weights do not have the same edge ids"
+                );
+                new_weights.0[i].insert(
+                    *self_id,
+                    self_ttf.apply(other_ttf, |w1, w2| {
+                        Time(w1.0.powf(a / (a + b)) * w2.0.powf(b / (a + b)))
+                    }),
+                );
             }
         }
         new_weights
@@ -113,7 +133,7 @@ impl<T: TTFNum> RoadNetworkWeights<T> {
 }
 
 impl<T> Index<usize> for RoadNetworkWeights<T> {
-    type Output = Vec<TTF<Time<T>>>;
+    type Output = HashMap<OriginalEdgeIndex, TTF<Time<T>>>;
     fn index(&self, x: usize) -> &Self::Output {
         &self.0[x]
     }
@@ -125,10 +145,10 @@ impl<T> IndexMut<usize> for RoadNetworkWeights<T> {
     }
 }
 
-impl<T> Index<(usize, EdgeIndex)> for RoadNetworkWeights<T> {
+impl<T> Index<(usize, OriginalEdgeIndex)> for RoadNetworkWeights<T> {
     type Output = TTF<Time<T>>;
-    fn index(&self, (vehicle_id, edge): (usize, EdgeIndex)) -> &Self::Output {
-        &self.0[vehicle_id][edge.index()]
+    fn index(&self, (vehicle_id, edge_id): (usize, OriginalEdgeIndex)) -> &Self::Output {
+        &self.0[vehicle_id][&edge_id]
     }
 }
 
@@ -138,19 +158,21 @@ mod tests {
 
     #[test]
     fn average_test() {
-        let w0 = RoadNetworkWeights(vec![vec![TTF::Constant(Time(10.))]]);
-        let w1 = RoadNetworkWeights(vec![vec![TTF::Constant(Time(20.))]]);
+        let w0 = RoadNetworkWeights(vec![[(0, TTF::Constant(Time(10.)))].into_iter().collect()]);
+        let w1 = RoadNetworkWeights(vec![[(0, TTF::Constant(Time(20.)))].into_iter().collect()]);
         // Result = 0.2 * 10 + 0.8 * 20 = 2 + 16 = 18.
-        let w2 = RoadNetworkWeights(vec![vec![TTF::Constant(Time(18.))]]);
+        let w2 = RoadNetworkWeights(vec![[(0, TTF::Constant(Time(18.)))].into_iter().collect()]);
         assert_eq!(w0.average(&w1, 0.2), w2);
     }
 
     #[test]
     fn genetic_average_test() {
-        let w0 = RoadNetworkWeights(vec![vec![TTF::Constant(Time(2.))]]);
-        let w1 = RoadNetworkWeights(vec![vec![TTF::Constant(Time(3.))]]);
+        let w0 = RoadNetworkWeights(vec![[(0, TTF::Constant(Time(2.)))].into_iter().collect()]);
+        let w1 = RoadNetworkWeights(vec![[(0, TTF::Constant(Time(3.)))].into_iter().collect()]);
         // Result = (2^3 * 3^2)^(1/5) = 72^(1/5).
-        let w2 = RoadNetworkWeights(vec![vec![TTF::Constant(Time(72.0f64.powf(0.2)))]]);
+        let w2 = RoadNetworkWeights(vec![[(0, TTF::Constant(Time(72.0f64.powf(0.2))))]
+            .into_iter()
+            .collect()]);
         assert_eq!(w0.genetic_average(&w1, 3.0, 2.0), w2);
     }
 }