+++ tests/bottleneck.rs
@@ -3,7 +3,7 @@
 // Licensed under the Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International
 // https://creativecommons.org/licenses/by-nc-nd/4.0/legalcode
 
-use hashbrown::{HashMap, HashSet};
+use hashbrown::HashSet;
 use metropolis::agent::Agent;
 use metropolis::learning::{ExponentialLearningModel, LearningModel};
 use metropolis::mode::trip::{DepartureTimeModel, Leg, LegType, RoadLeg, TravelingMode};
@@ -19,7 +19,6 @@ use metropolis::simulation::Simulation;
 use metropolis::stop::StopCriterion;
 use metropolis::travel_utility::TravelUtility;
 use metropolis::units::{Flow, Interval, Length, Speed, Time, PCE};
-use petgraph::graph::{edge_index, node_index, DiGraph};
 use ttf::TTF;
 
 fn get_simulation(overtaking: bool) -> Simulation<f64> {
@@ -48,11 +47,12 @@ fn get_simulation(overtaking: bool) -> Simulation<f64> {
     // Create a road network with 2 edges, with capacities 1 vehicle each 2 seconds and 1 vehicle
     // each 4 seconds, respectively.
     // Travel time is 1 second free-flow.
-    let graph = DiGraph::from_edges(&[
+    let edges = vec![
         (
             0,
             1,
             RoadEdge::new(
+                0,
                 Speed(1.0),
                 Length(1.0),
                 1,
@@ -66,6 +66,7 @@ fn get_simulation(overtaking: bool) -> Simulation<f64> {
             1,
             2,
             RoadEdge::new(
+                1,
                 Speed(1.0),
                 Length(1.0),
                 1,
@@ -75,11 +76,7 @@ fn get_simulation(overtaking: bool) -> Simulation<f64> {
                 overtaking,
             ),
         ),
-    ]);
-    let node_map: HashMap<_, _> = (0..=2)
-        .into_iter()
-        .map(|i| (i as u64, node_index(i)))
-        .collect();
+    ];
     let vehicle = Vehicle::new(
         Length(1.0),
         PCE(1.0),
@@ -87,7 +84,7 @@ fn get_simulation(overtaking: bool) -> Simulation<f64> {
         HashSet::new(),
         HashSet::new(),
     );
-    let road_network = RoadNetwork::new(graph, node_map, vec![vehicle]);
+    let road_network = RoadNetwork::from_edges(edges, vec![vehicle]);
     let network = Network::new(Some(road_network));
 
     let parameters = Parameters {
@@ -183,7 +180,7 @@ fn bottleneck_no_overtaking_test() {
         .network_weights()
         .get_road_network()
         .unwrap();
-    let edge_weight = &weights[(0, edge_index(0))];
+    let edge_weight = &weights[(0, 0)];
     let TTF::Piecewise(ttf) = edge_weight else {
         panic!("TTF should be piecewise");
     };
@@ -264,7 +261,7 @@ fn bottleneck_overtaking_test() {
         .network_weights()
         .get_road_network()
         .unwrap();
-    let edge_weight = &weights[(0, edge_index(0))];
+    let edge_weight = &weights[(0, 0)];
     let TTF::Piecewise(ttf) = edge_weight else {
         panic!("TTF should be piecewise");
     };